refactor: split sosotalk post detail into header, body, and actions
