+++ src/app/sosotalk/[postId]/_components/sosotalk-post-detail/sosotalk-post-detail.types.ts
@@ -17,3 +17,30 @@ export interface SosoTalkPostDetailProps {
   onDeleteClick?: () => void;
   onShareClick?: () => void;
 }
+
+export interface SosoTalkPostHeaderProps {
+  title: string;
+  authorName: string;
+  authorImageUrl?: string;
+  categoryLabel?: string;
+  statusLabel?: string;
+  createdAt: string;
+  createdAtDateTime?: string;
+  isAuthor?: boolean;
+  onMoreClick?: () => void;
+  onEditClick?: () => void;
+  onDeleteClick?: () => void;
+}
+
+export interface SosoTalkPostBodyProps {
+  title: string;
+  content: string;
+  imageUrl?: string;
+}
+
+export interface SosoTalkPostActionsProps {
+  contentCharacterCount?: number;
+  likeCount?: number;
+  commentCount?: number;
+  onShareClick?: () => void;
+}
